+++ src/external/keep3r/Keep3rV1Oracle.sol
@@ -3,12 +3,18 @@ pragma solidity 0.6.12;
 
 abstract contract Keep3rV1Oracle {
     struct Observation {
-        uint timestamp;
-        uint price0Cumulative;
-        uint price1Cumulative;
+        uint256 timestamp;
+        uint256 price0Cumulative;
+        uint256 price1Cumulative;
     }
 
     function factory() external pure virtual returns (address);
+
     mapping(address => Observation[]) public observations;
-    function observationLength(address pair) external view virtual returns (uint);
+
+    function observationLength(address pair)
+        external
+        view
+        virtual
+        returns (uint256);
 }
